extract out ReflectionHelper & begin abstraction
