+++ src/main/java/me/rayzr522/jsonmessage/compat/ChatPacket.java
@@ -1,0 +1,6 @@
+package me.rayzr522.jsonmessage.compat;
+
+public interface ChatPacket {
+    Object createActionbarPacket(Object chatComponent);
+    Object createTextPacket(Object chatComponent);
+}
