+++ src/main/java/me/rayzr522/jsonmessage/compat/PlayerConnection.java
@@ -1,0 +1,7 @@
+package me.rayzr522.jsonmessage.compat;
+
+import org.bukkit.entity.Player;
+
+public interface PlayerConnection {
+    void sendPacket(Object packet, Player... players);
+}
