+++ src/main/java/me/rayzr522/jsonmessage/compat/TitlePacket.java
@@ -1,0 +1,12 @@
+package me.rayzr522.jsonmessage.compat;
+
+public interface TitlePacket {
+
+    Object createTitlePacket(String message);
+
+    Object createTitleTimesPacket(int fadeIn, int stay, int fadeOut);
+
+    Object createSubtitlePacket(Object chatComponent);
+
+    Object createSubtitleTimesPacket(int fadeIn, int stay, int fadeOut);
+}
